% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timescape.R
\name{checkAlpha}
\alias{checkAlpha}
\title{check alpha value input is correct}
\usage{
checkAlpha(alpha)
}
\arguments{
\item{alpha}{-- alpha provided by user}
}
\value{
None
}
\description{
check alpha value input is correct
}
\examples{
checkAlpha(4)
checkAlpha(100)
}

