% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timescape.R
\name{checkClonalPrev}
\alias{checkClonalPrev}
\title{check clonal_prev parameter data}
\usage{
checkClonalPrev(clonal_prev)
}
\arguments{
\item{clonal_prev}{-- clonal prevalence provided by user}
}
\description{
check clonal_prev parameter data
}

