% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timescape.R
\name{checkCloneColours}
\alias{checkCloneColours}
\title{check clone_colours parameter}
\usage{
checkCloneColours(clone_colours)
}
\arguments{
\item{clone_colours}{-- clone_colours provided by user}
}
\description{
check clone_colours parameter
}

