% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timesweep.R
\name{checkGtypePositioning}
\alias{checkGtypePositioning}
\title{check genotype_position parameter}
\usage{
checkGtypePositioning(genotype_position)
}
\arguments{
\item{genotype_position}{-- genotype_position provided by user}
}
\description{
check genotype_position parameter
}

