% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timescape.R
\name{checkMinDims}
\alias{checkMinDims}
\title{Function to check minimum dimensions}
\usage{
checkMinDims(mutations, height, width)
}
\arguments{
\item{mutations}{-- mutations provided by user}

\item{height}{-- height provided by user}

\item{width}{-- width provided by user}
}
\value{
None
}
\description{
Function to check minimum dimensions
}
\examples{
checkMinDims(data.frame(chr = c("11"), coord = c(104043), VAF = c(0.1)), "700px", "700px")
}

