% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timesweep.R
\name{checkPerts}
\alias{checkPerts}
\title{check perturbations parameter}
\usage{
checkPerts(perturbations)
}
\arguments{
\item{perturbations}{-- perturbations provided by user}
}
\description{
check perturbations parameter
}

