% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timescape.R
\name{checkRequiredInputs}
\alias{checkRequiredInputs}
\title{Function to check required inputs are present}
\usage{
checkRequiredInputs(clonal_prev, tree_edges)
}
\arguments{
\item{clonal_prev}{-- clonal_prev provided by user}

\item{tree_edges}{-- tree_edges provided by user}
}
\value{
None
}
\description{
Function to check required inputs are present
}
\examples{
checkRequiredInputs(data.frame(timepoint = c(rep("Diagnosis", 6), rep("Relapse", 1)), clone_id = c("1","2","3","4","5","6","7"), clonal_prev = c("0.1","0.22","0.08","0.53","0.009","0.061","1")), 
data.frame(source = c("1","1","2","2","5","6"), target=c("2","5","3","4","6","7")))
checkRequiredInputs(data.frame(timepoint = c(rep("Diagnosis", 6), rep("Relapse", 1)), clone_id = c("1","2","3","4","5","6","7"), clonal_prev = c("0.12","0.12","0.18","0.13","0.009","0.061","1")), 
data.frame(source = c("1","1","2","2","5","6"), target=c("2","5","3","4","6","7")))
}

