% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timesweep.R
\name{checkSort}
\alias{checkSort}
\title{check sort genotypes parameter input is correct}
\usage{
checkSort(sort)
}
\arguments{
\item{sort}{-- sort param provided by user}
}
\description{
check sort genotypes parameter input is correct
}

