% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timesweep.R
\name{checkTreeEdges}
\alias{checkTreeEdges}
\title{check tree_edges parameter data}
\usage{
checkTreeEdges(tree_edges)
}
\arguments{
\item{tree_edges}{-- tree edges provided by user}
}
\description{
check tree_edges parameter data
}

