% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timesweep.R
\name{getMutationsData}
\alias{getMutationsData}
\title{get mutation data}
\usage{
getMutationsData(mutations, tree_edges, clonal_prev)
}
\arguments{
\item{mutations}{-- mutations data from user}

\item{tree_edges}{-- tree edges data from user}

\item{clonal_prev}{-- clonal prevalence data from user}
}
\description{
get mutation data
}

