% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timesweep.R
\name{processUserData}
\alias{processUserData}
\title{Function to process the user data}
\usage{
processUserData(clonal_prev, tree_edges, mutations, clone_colours, xaxis_title,
  yaxis_title, phylogeny_title, alpha, genotype_position, perturbations, sort,
  show_warnings, width, height)
}
\description{
Function to process the user data
}

