% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timescape.R
\name{renderTimescape}
\alias{renderTimescape}
\title{Widget render function for use in Shiny}
\usage{
renderTimescape(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{-- expression for Shiny}

\item{env}{-- environment for Shiny}

\item{quoted}{-- default is FALSE}
}
\value{
None
}
\description{
Widget render function for use in Shiny
}

