% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timescape.R
\name{replaceSpaces}
\alias{replaceSpaces}
\title{function to replace spaces with underscores in all data frames & keep maps of original names to space-replaced names}
\usage{
replaceSpaces(clonal_prev, tree_edges, clone_colours, mutation_info, mutations,
  mutation_prevalences)
}
\arguments{
\item{clonal_prev}{-- clonal_prev data from user}

\item{tree_edges}{-- tree edges data from user}

\item{clone_colours}{-- clone_colours data from user}

\item{mutation_info}{-- processed mutation_info}

\item{mutations}{-- mutations data from user}

\item{mutation_prevalences}{-- mutation_prevalences data from user}
}
\description{
function to replace spaces with underscores in all data frames & keep maps of original names to space-replaced names
}
\examples{
replaceSpaces(mutations = data.frame(chrom = c("11"), coord = c(104043), VAF = c(0.1), clone_id=c(1), timepoint=c("Relapse")), 
tree_edges = data.frame(source = c("1","1","2","2","5","6"), target=c("2","5","3","4","6","7")), 
clonal_prev = data.frame(timepoint = c(rep("Diagnosis", 6), rep("Relapse", 1)), clone_id = c("1","2","3","4","5","6","7"), clonal_prev = c("0.12","0.12","0.18","0.13","0.009","0.061","1")),
mutation_prevalences = list("X:6154028" = data.frame(timepoint = c("Diagnosis"), VAF = c(0.5557))), mutation_info=data.frame(clone_id=c(1)),
clone_colours = data.frame(clone_id = c("1","2","3", "4"), colour = c("#beaed4", "#fdc086", "#beaed4", "#beaed4")))
}

