% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timescape.R
\name{timescape}
\alias{timescape}
\title{TimeScape}
\usage{
timescape(clonal_prev, tree_edges, mutations = "NA", clone_colours = "NA",
  xaxis_title = "Time Point", yaxis_title = "Clonal Prevalence",
  phylogeny_title = "Clonal Phylogeny", alpha = 50,
  genotype_position = "stack", perturbations = "NA", sort = FALSE,
  show_warnings = TRUE, width = 900, height = NULL)
}
\arguments{
\item{clonal_prev}{(Data Frame) Clonal prevalence. Note: timepoints will be alphanumerically sorted in the view.
Format: columns are (1) character() "timepoint" - time point
                    (2) character() "clone_id" - clone id
                    (3) numeric() "clonal_prev" - clonal prevalence.}

\item{tree_edges}{(Data Frame) Tree edges of a rooted tree.
Format: columns are (1) character() "source" - source node id
                    (2) character() "target" - target node id.}

\item{mutations}{(Data Frame) (Optional) Mutations occurring at each clone. Any additional field will be shown in the mutation table.
Format: columns are (1) character() "chrom" - chromosome number
                    (2) numeric() "coord" - coordinate of mutation on chromosome
                    (3) character() "clone_id" - clone id
                    (4) character() "timepoint" - time point
                    (5) numeric() "VAF" - variant allele frequency of the mutation in the corresponding timepoint.}

\item{clone_colours}{(Data Frame) (Optional) Clone ids and their corresponding colours 
Format: columns are (1) character() "clone_id" - the clone ids
                    (2) character() "colour" - the corresponding Hex colour for each clone id.}

\item{xaxis_title}{character() (Optional) x-axis title. Default is "Time Point".}

\item{yaxis_title}{character() (Optional) y-axis title. Default is "Clonal Prevalence".}

\item{phylogeny_title}{character() (Optional) Legend phylogeny title. Default is "Clonal Phylogeny".}

\item{alpha}{numeric() (Optional) Alpha value for sweeps, range [0, 100].}

\item{genotype_position}{character() (Optional) How to position the genotypes from ["centre", "stack", "space"] 
"centre" -- genotypes are centred with respect to their ancestors
"stack" -- genotypes are stacked such that no genotype is split at any time point
"space" -- genotypes are stacked but with a bit of spacing at the bottom}

\item{perturbations}{(Data Frame) (Optional) Any perturbations that occurred between two time points, 
and the fraction of total tumour content remaining.
Format: columns are (1) character() "pert_name" - the perturbation name
                    (2) character() "prev_tp" - the time point (as labelled in clonal prevalence data) 
                                             BEFORE perturbation
                    (3) numeric() "frac" - the fraction of total tumour content remaining at the 
                                          time of perturbation, range [0, 1].}

\item{sort}{logical() (Optional) Whether (TRUE) or not (FALSE) to vertically sort the genotypes by their emergence values (descending). 
Default is FALSE. 
Note that genotype sorting will always retain the phylogenetic hierarchy, and this parameter will only affect the ordering of siblings.}

\item{show_warnings}{logical() (Optional) Whether or not to show any warnings. Default is TRUE.}

\item{width}{numeric() (Optional) Width of the plot. Minimum width is 450.}

\item{height}{numeric() (Optional) Height of the plot. Minimum height with and without mutations is 500 and 260, respectively.}
}
\value{
None
}
\description{
\code{timescape} is a tool for visualizing temporal clonal evolution data.
}
\details{
Interactive components: \cr

1. Mouseover any clone to view its (i) clone ID and (ii) clonal prevalence at each time point. \cr
2. Click the view switch button to switch from the traditional timescape view to the clonal trajectory view, where each clone changes prevalence on its own track. \cr
3. Click the download buttons to download a PNG or SVG of the view. \cr
}
\examples{

# EXAMPLE 1 - Acute myeloid leukemia patient, Ding et al., 2012

# genotype tree edges
tree_edges <- read.csv(system.file("extdata", "AML_tree_edges.csv", package = "timescape"))
# clonal prevalences
clonal_prev <- read.csv(system.file("extdata", "AML_clonal_prev.csv", package = "timescape"))
# targeted mutations
mutations <- read.csv(system.file("extdata", "AML_mutations.csv", package = "timescape"))
# perturbations
perturbations <- data.frame( pert_name = c("Chemotherapy"), 
                             prev_tp = c("Diagnosis"),
                              frac = c(0.1))
# run timescape
timescape(clonal_prev = clonal_prev, tree_edges = tree_edges, perturbations = perturbations, mutations = mutations)

# EXAMPLE 2 - Patient 7, McPherson and Roth et al., 2016


# genotype tree edges
tree_edges <- read.csv(system.file("extdata", "px7_tree_edges.csv", package = "timescape"))
# clonal prevalences
clonal_prev <- read.csv(system.file("extdata", "px7_clonal_prev.csv", package = "timescape"))
# clone colours
clone_colours <- data.frame(clone_id = c("A","B","C","D","E"), 
                            colour = c("d0ced0", "2CD0AB", "FFD94B", "FD8EE5", "F8766D"))
# run timescape
timescape(clonal_prev = clonal_prev, tree_edges = tree_edges, clone_colours = clone_colours, height=260, alpha=15)
}

