% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timescape.R
\name{timescapeOutput}
\alias{timescapeOutput}
\title{Widget output function for use in Shiny}
\usage{
timescapeOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{-- id of output}

\item{width}{-- width of output}

\item{height}{-- height of output}
}
\value{
None
}
\description{
Widget output function for use in Shiny
}
\examples{
timescapeOutput(1, '100\%', '300px')
timescapeOutput(1, '80\%', '300px')
}

