% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timescape.R
\name{checkCloneColours}
\alias{checkCloneColours}
\title{check clone_colours parameter}
\usage{
checkCloneColours(clone_colours)
}
\arguments{
\item{clone_colours}{-- clone_colours provided by user}
}
\value{
None
}
\description{
check clone_colours parameter
}
\examples{
checkCloneColours(data.frame(clone_id = c("1","2","3", "4"), colour = c("#beaed4", "#fdc086", "#beaed4", "#beaed4")))
}

