% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timesweep.R
\name{checkMinDims}
\alias{checkMinDims}
\title{Function to check minimum dimensions}
\usage{
checkMinDims(mutations, height, width)
}
\arguments{
\item{mutations}{-- mutations provided by user}

\item{height}{-- height provided by user}

\item{width}{-- width provided by user}
}
\description{
Function to check minimum dimensions
}

