% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timescape.R
\name{checkPerts}
\alias{checkPerts}
\title{check perturbations parameter}
\usage{
checkPerts(perturbations)
}
\arguments{
\item{perturbations}{-- perturbations provided by user}
}
\value{
Perturbations after checking them for content types and column names
}
\description{
check perturbations parameter
}
\examples{
checkPerts(data.frame(pert_name = c("New Drug"), prev_tp = c("Diagnosis"), frac = c(0.1)))
}

