% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timesweep.R
\name{checkRequiredInputs}
\alias{checkRequiredInputs}
\title{Function to check required inputs are present}
\usage{
checkRequiredInputs(clonal_prev, tree_edges)
}
\arguments{
\item{clonal_prev}{-- clonal_prev provided by user}

\item{tree_edges}{-- tree_edges provided by user}
}
\description{
Function to check required inputs are present
}

