% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timescape.R
\name{checkTreeEdges}
\alias{checkTreeEdges}
\title{check tree_edges parameter data}
\usage{
checkTreeEdges(tree_edges)
}
\arguments{
\item{tree_edges}{-- tree edges provided by user}
}
\description{
check tree_edges parameter data
}
\examples{
checkTreeEdges(data.frame(source = c("1","1","2","2","5","6"), target=c("2","5","3","4","6","7")))
}

