% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timesweep.R
\name{replaceSpaces}
\alias{replaceSpaces}
\title{function to replace spaces with underscores in all data frames & keep maps of original names to space-replaced names}
\usage{
replaceSpaces(clonal_prev, tree_edges, clone_colours, mutation_info, mutations,
  mutation_prevalences)
}
\arguments{
\item{clonal_prev}{-- clonal_prev data from user}

\item{tree_edges}{-- tree edges data from user}

\item{clone_colours}{-- clone_colours data from user}

\item{mutation_info}{-- processed mutation_info}

\item{mutations}{-- mutations data from user}

\item{mutation_prevalences}{-- mutation_prevalences data from user}
}
\description{
function to replace spaces with underscores in all data frames & keep maps of original names to space-replaced names
}

