% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timescape.R
\name{timescape}
\alias{timescape}
\title{Timecapes}
\usage{
timescape(clonal_prev, tree_edges, mutations = "NA", clone_colours = "NA",
  xaxis_title = "Time Point", yaxis_title = "Clonal Prevalence",
  phylogeny_title = "Clonal Phylogeny", alpha = 50,
  genotype_position = "stack", perturbations = "NA", sort = FALSE,
  show_warnings = TRUE, width = 900, height = NULL)
}
\arguments{
\item{clonal_prev}{{Data Frame} Clonal prevalence. Note: timepoints will be alphanumerically sorted in the view.
Format: columns are (1) {String} "timepoint" - time point
                    (2) {String} "clone_id" - clone id
                    (3) {Number} "clonal_prev" - clonal prevalence.}

\item{tree_edges}{{Data Frame} Tree edges of a rooted tree.
Format: columns are (1) {String} "source" - source node id
                    (2) {String} "target" - target node id.}

\item{mutations}{{Data Frame} (Optional) Mutations occurring at each clone. Any additional field will be shown in the mutation table.
Format: columns are (1) {String} "chrom" - chromosome number
                    (2) {Number} "coord" - coordinate of mutation on chromosome
                    (3) {String} "clone_id" - clone id
                    (4) {String} "timepoint" - time point
                    (5) {Number} "VAF" - variant allele frequency of the mutation in the corresponding sample.}

\item{clone_colours}{{Data Frame} (Optional) Clone ids and their corresponding colours 
Format: columns are (1) {String} "clone_id" - the clone ids
                    (2) {String} "colour" - the corresponding Hex colour for each clone id.}

\item{xaxis_title}{{String} (Optional) x-axis title. Default is "Time Point".}

\item{yaxis_title}{{String} (Optional) y-axis title. Default is "Clonal Prevalence".}

\item{phylogeny_title}{{String} (Optional) Legend phylogeny title. Default is "Clonal Phylogeny".}

\item{alpha}{{Number} (Optional) Alpha value for sweeps, range [0, 100].}

\item{genotype_position}{{String} (Optional) How to position the genotypes from ["centre", "stack", "space"] 
"centre" -- genotypes are centred with respect to their ancestors
"stack" -- genotypes are stacked such that no genotype is split at any time point
"space" -- genotypes are stacked but with a bit of spacing at the top (emergence is clearer)}

\item{perturbations}{{Data Frame} (Optional) Any perturbations that occurred between two time points, 
and the fraction of total tumour content left.
Format: columns are (1) {String} "pert_name" - the perturbation name
                    (2) {String} "prev_tp" - the time point (as labelled in clonal prevalence data) 
                                             BEFORE perturbation
                    (3) {Number} "frac" - the fraction of total tumour content remaining at the 
                                          time of perturbation, range [0, 1].}

\item{sort}{{Boolean} (Optional) Whether (TRUE) or not (FALSE) to vertically sort the genotypes by their emergence values (descending).}

\item{show_warnings}{{Boolean} (Optional) Whether or not to show any warnings. Default is TRUE.}

\item{width}{{Number} (Optional) Width of the plot. Minimum width is 450.}

\item{height}{{Number} (Optional) Height of the plot. Minimum height with and without mutations is 500 and 260, respectively.}
}
\description{
\code{timescape} generates patient clonal timescapes.
}
\examples{
library("timescape")

# clonal prevalences
clonal_prev <- data.frame( timepoint = c(rep("T1", 6), rep("T2", 6)), 
                           clone_id = c("1","6","5","4","3","2","1","6","5","4","3","2"),
                           clonal_prev = c("0.0205127","0.284957","0.637239","0.0477972","0.00404099","0.00545235",
                                           "0.0134362","0.00000150677","0.00000385311","0.000627522","0.551521","0.43441"))
# genotype tree edges
tree_edges <- data.frame(source = c("1","1","6","5","3"), 
                         target = c("3","6","5","4","2"))
# clone colours
clone_colours <- data.frame( clone_id = c("1","2","3","4","5","6"), 
                             colour = c("F8766D66", "B79F0066", "00BA3866", "00BFC466", "619CFF66", "F564E366"))
# perturbations (chemotherapy)
perturbations <- data.frame( pert_name = c("Chemo"), 
                             prev_tp = c("T1"),
                              frac = c(0.1))
# run timescape
timescape(clonal_prev = clonal_prev, tree_edges = tree_edges, clone_colours = clone_colours, perturbations = perturbations, height=400)
}

