% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timesweep.R
\name{timesweep}
\alias{timesweep}
\title{Timesweeps}
\usage{
timesweep(patient, clonal.prev.csv, tree.gml, node.col, width = NULL,
  height = NULL)
}
\arguments{
\item{patient}{Patient name.}

\item{clonal.prev.csv}{Path to clonal prevalence csv file.}

\item{tree.gml}{Path to GML file.}

\item{node.col}{Named character vector with the keys as node labels and 
values as colors.}

\item{width}{Width of the plot.}

\item{height}{Height of the plot.}
}
\description{
\code{timesweep} generates patient clonal timesweeps.
}

