% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timesweep.R
\name{timesweep}
\alias{timesweep}
\title{Timesweeps}
\usage{
timesweep(clonal.prev, tree.edges, node.col = "NA",
  xaxis.title = "Time Point", yaxis.title = "Relative Cellular Prevalence",
  alpha = 30, genotype.position = "stack", show.root = "T",
  perturbations = "NA", sort = "T", width = NULL, height = NULL)
}
\arguments{
\item{clonal.prev}{Clonal prevalence data frame.
Format: columns are (1) {String} patient name
                    (2) {String} time point
                    (3) {String} clone id
                    (4) {Number} clonal prevalence.}

\item{node.col}{Data frame with node labels and colours 
Format: columns are (1) {String} the node labels
                    (2) {String} the corresponding colour for each node label.
e.g. data.frame(node_label = c("1","2","3","4","5"), 
                 col = c("F8766D66", "A3A50066", "00BF7D66", "00B0F666", "E76BF366"))}

\item{xaxis.title}{x-axis title.}

\item{yaxis.title}{y-axis title.}

\item{alpha}{Alpha value for sweeps}

\item{genotype.position}{How to position the genotypes from ["centre", "stack", "space"] 
"centre" -- genotypes are centred with respect to their ancestors
"stack" -- genotypes are stacked such that no genotype is split at any time point
"space" -- genotypes are stacked but with a bit of spacing at the top (emergence is clearer)}

\item{show.root}{Whether or not to show the root in the timesweep view}

\item{perturbations}{Data frame of any perturbations that occurred between two time points, 
and the fraction of total tumour content left.
Format: columns are (1) {String} the perturbation name
                    (2) {String} the time point (as labelled in cellular prevalence data) BEFORE perturbation
                    (3) {String} the time point (as labelled in cellular prevalence data) AFTER perturbation
                    (4) {Number} the fraction of total tumour content remaining at the time of perturbation.
e.g. data.frame(perturbation = c("Chemo"), 
                 prev_tp = c("T1"),
                 next_tp = c("T2"), 
                 frac = c(0.2))}

\item{sort}{Whether (T) or not (F) to vertically sort the genotypes by their emergence values}

\item{width}{Width of the plot.}

\item{height}{Height of the plot.}

\item{tree}{Tree edges data frame.
Format: columns are (1) {String} patient name
                    (2) {String} source node id
                    (3) {String} target node id.
e.g. data.frame(patient_name = c("SAMPLE_PATIENT"), 
                source = c("Root","1","1","6","5","3"), 
                target=c("1","3","6","5","4","2"))}
}
\description{
\code{timesweep} generates patient clonal timesweeps.
}
\examples{
library("timesweep")
clonal.prev <- read.csv(system.file("extdata", "clonal_dynamics.csv", package = "timesweep"))
tree.edges <- data.frame(patient_name = c("SAMPLE_PATIENT"), source = c("Root","1","1","6","5","3"), target=c("1","3","6","5","4","2"))
timesweep(clonal.prev = clonal.prev, tree.edges = tree.edges)
}

