% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timesweep.R
\name{timesweep}
\alias{timesweep}
\title{Timesweeps}
\usage{
timesweep(clonal_prev, tree_edges, clone_colours = "NA",
  xaxis_title = "Time Point", yaxis_title = "Relative Cellular Prevalence",
  alpha = 50, genotype_position = "stack", show_root = TRUE,
  perturbations = "NA", sort = TRUE, width = NULL, height = NULL)
}
\arguments{
\item{clonal_prev}{Clonal prevalence data frame.
Format: columns are (1) {String} "timepoint" - time point
                    (2) {String} "clone_id" - clone id
                    (3) {Number} "clonal_prev" - clonal prevalence.}

\item{tree_edges}{Tree edges data frame. The root of the tree (id: "Root") must be specified as a source.
Format: columns are (1) {String} "source" - source node id
                    (2) {String} "target" - target node id.}

\item{clone_colours}{Data frame with clone ids and their corresponding colours 
Format: columns are (1) {String} "clone_id" - the clone ids
                    (2) {String} "colour" - the corresponding Hex colour for each clone id.}

\item{xaxis_title}{x-axis title.}

\item{yaxis_title}{y-axis title.}

\item{alpha}{Alpha value for sweeps, range [0, 100].}

\item{genotype_position}{How to position the genotypes from ["centre", "stack", "space"] 
"centre" -- genotypes are centred with respect to their ancestors
"stack" -- genotypes are stacked such that no genotype is split at any time point
"space" -- genotypes are stacked but with a bit of spacing at the top (emergence is clearer)}

\item{show_root}{Whether (TRUE) or not (FALSE) to show the root in the timesweep view.}

\item{perturbations}{Data frame of any perturbations that occurred between two time points, 
and the fraction of total tumour content left.
Format: columns are (1) {String} "pert_name" - the perturbation name
                    (2) {String} "prev_tp" - the time point (as labelled in clonal prevalence data) 
                                             BEFORE perturbation
                    (3) {Number} "frac" - the fraction of total tumour content remaining at the 
                                          time of perturbation, range [0, 1].}

\item{sort}{Whether (TRUE) or not (FALSE) to vertically sort the genotypes by their emergence values (descending).}

\item{width}{Width of the plot.}

\item{height}{Height of the plot.}
}
\description{
\code{timesweep} generates patient clonal timesweeps.
}
\examples{
library("timesweep")
clonal_prev <- data.frame( timepoint = c(rep("T1", 6), rep("T2", 6)), 
                           clone_id = c("1","6","5","4","3","2","1","6","5","4","3","2"),
                           clonal_prev = c("0.0205127","0.284957","0.637239","0.0477972","0.00404099","0.00545235",
                                           "0.0134362","0.00000150677","0.00000385311","0.000627522","0.551521","0.43441"))
tree_edges <- data.frame(source = c("Root","1","1","6","5","3"), 
                         target = c("1","3","6","5","4","2"))
clone_colours <- data.frame( clone_id = c("1","2","3","4","5","6"), 
                             colour = c("F8766D66", "B79F0066", "00BA3866", "00BFC466", "619CFF66", "F564E366"))
perturbations <- data.frame( pert_name = c("Chemo"), 
                             prev_tp = c("T1"),
                              frac = c(0.1))
timesweep(clonal_prev = clonal_prev, tree_edges = tree_edges, clone_colours = clone_colours, perturbations = perturbations)
}

